+++ src/test/java/com/powsybl/openloadflow/network/VoltageTargetPriorityTest.java
@@ -7,12 +7,22 @@
  */
 package com.powsybl.openloadflow.network;
 
+import com.powsybl.iidm.network.Bus;
+import com.powsybl.iidm.network.Network;
+import com.powsybl.iidm.network.test.EurostagTutorialExample1Factory;
+import com.powsybl.loadflow.LoadFlow;
+import com.powsybl.loadflow.LoadFlowParameters;
+import com.powsybl.loadflow.LoadFlowResult;
+import com.powsybl.math.matrix.DenseMatrixFactory;
+import com.powsybl.openloadflow.OpenLoadFlowParameters;
+import com.powsybl.openloadflow.OpenLoadFlowProvider;
 import org.junit.jupiter.api.Test;
 
 import java.util.List;
 
-import static org.junit.jupiter.api.Assertions.assertEquals;
-import static org.junit.jupiter.api.Assertions.assertThrows;
+import static com.powsybl.openloadflow.util.LoadFlowAssert.assertVoltageEquals;
+import static org.junit.jupiter.api.Assertions.*;
+import static org.junit.jupiter.api.Assertions.assertTrue;
 
 /**
  * @author Pierre Arvy {@literal <pierre.arvy at artelys.com>}
@@ -34,6 +44,59 @@ class VoltageTargetPriorityTest {
 
         List<Integer> voltageTargetPriority2 = List.of(0, 1, 4);
         assertThrows(IllegalStateException.class, () -> VoltageControl.setVoltageTargetPriority(voltageTargetPriority2));
+    }
+
+    @Test
+    void voltageTargetPriority() {
+        Network network = EurostagFactory.fix(EurostagTutorialExample1Factory.create());
+        Bus genBus = network.getBusBreakerView().getBus("NGEN");
+        Bus loadBus = network.getBusBreakerView().getBus("NLOAD");
+
+        LoadFlow.Runner loadFlowRunner = new LoadFlow.Runner(new OpenLoadFlowProvider(new DenseMatrixFactory()));
+        LoadFlowParameters parameters = new LoadFlowParameters().setUseReactiveLimits(false)
+                .setDistributedSlack(false);
+        OpenLoadFlowParameters parametersExt = OpenLoadFlowParameters.create(parameters)
+                .setSlackBusSelectionMode(SlackBusSelectionMode.FIRST);
+
+        network.getTwoWindingsTransformer("NHV2_NLOAD")
+                .getRatioTapChanger()
+                .setRegulationTerminal(network.getTwoWindingsTransformer("NGEN_NHV1").getTerminal1())
+                .setTargetV(25.115);
+
+        loadBus.getVoltageLevel().newShuntCompensator()
+                .setId("SC")
+                .setBus(loadBus.getId())
+                .setConnectableBus(loadBus.getId())
+                .setSectionCount(1)
+                .newLinearModel()
+                .setBPerSection(3.25 * Math.pow(10, -3))
+                .setMaximumSectionCount(1)
+                .add()
+                .setVoltageRegulatorOn(true)
+                .setRegulatingTerminal(network.getTwoWindingsTransformer("NGEN_NHV1").getTerminal1())
+                .setTargetV(23.75)
+                .setTargetDeadband(0)
+                .add();
+
+        parameters.setTransformerVoltageControlOn(true)
+                .setShuntCompensatorVoltageControlOn(true);
+
+        // Generator has target priority by default
+        LoadFlowResult result = loadFlowRunner.run(network, parameters);
+        assertTrue(result.isFullyConverged());
+        assertVoltageEquals(24.5, genBus);
+
+        // Transformer has target priority
+        parametersExt.setVoltageTargetPriority(List.of(ControlTargetPriority.TRANSFORMER.name(), ControlTargetPriority.GENERATOR.name(), ControlTargetPriority.SHUNT.name()));
+        result = loadFlowRunner.run(network, parameters);
+        assertTrue(result.isFullyConverged());
+        assertVoltageEquals(25.115, genBus);
+
+        // Shunt has target priority
+        parametersExt.setVoltageTargetPriority(List.of(ControlTargetPriority.SHUNT.name(), ControlTargetPriority.GENERATOR.name(), ControlTargetPriority.TRANSFORMER.name()));
+        result = loadFlowRunner.run(network, parameters);
+        assertTrue(result.isFullyConverged());
+        assertVoltageEquals(23.75, genBus);
     }
 
 }
